requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.11.18
aiosignal==1.3.2
arabic-reshaper==3.0.0
asn1crypto==1.5.1
attrs==25.3.0
audioop-lts==0.2.1
Brotli==1.1.0
certifi==2025.4.26
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.4.1
click==8.1.8
contourpy==1.3.2
cryptography==44.0.2
cssselect2==0.8.0
cycler==0.12.1
defusedxml==0.7.1
discord.py==2.5.2
distlib==0.3.9
distro==1.7.0
dotenv==0.9.9
filelock==3.18.0
fonttools==4.57.0
fpdf==1.7.2
fpdf2==2.8.3
frozenlist==1.6.0
greenlet==3.2.1
html5lib==1.1
idna==3.10
kiwisolver==1.4.8
lxml==5.4.0
matplotlib==3.10.1
multidict==6.4.3
numpy==2.2.5
oscrypto==1.3.0
packaging==25.0
pandas==2.2.3
pillow==11.2.1
pipenv==2025.0.1
platformdirs==4.3.7
playwright==1.52.0
propcache==0.3.1
pycparser==2.22
pydyf==0.11.0
pyee==13.0.0
pyHanko==0.26.0
pyhanko-certvalidator==0.26.8
pyparsing==3.2.3
pypdf==5.4.0
pyphen==0.17.2
python-bidi==0.6.6
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
pytz==2025.2
PyYAML==6.0.2
qrcode==8.1
reportlab==4.4.0
requests==2.32.3
setuptools==80.3.0
six==1.17.0
ssh-import-id==5.11
supervisor==4.2.1
svglib==1.5.1
tinycss2==1.4.0
tinyhtml5==2.0.0
typing_extensions==4.13.2
tzdata==2025.2
tzlocal==5.3.1
uritools==4.0.3
urllib3==2.4.0
virtualenv==20.30.0
weasyprint==65.1
webencodings==0.5.1
xhtml2pdf==0.2.17
yarl==1.20.0
zopfli==0.2.3.post1
